Relocated configs into another package
