+++ src/main/java/kg/peaksoft/ebookb4/api/AuthController.java
@@ -4,8 +4,8 @@ import kg.peaksoft.ebookb4.dto.request.LoginRequest;
 import kg.peaksoft.ebookb4.dto.request.SignupRequestClient;
 import kg.peaksoft.ebookb4.dto.request.SignupRequestVendor;
 import kg.peaksoft.ebookb4.dto.response.JwtResponse;
-import kg.peaksoft.ebookb4.security.jwt.JwtUtils;
-import kg.peaksoft.ebookb4.security.services.UserDetailsImpl;
+import kg.peaksoft.ebookb4.config.jwt.JwtUtils;
+import kg.peaksoft.ebookb4.config.services.UserDetailsImpl;
 import kg.peaksoft.ebookb4.db.service.ClientService;
 import kg.peaksoft.ebookb4.db.service.VendorService;
 import lombok.AllArgsConstructor;
