+++ src/main/java/kg/peaksoft/ebookb4/config/jwt/AuthTokenFilter.java
@@ -1,6 +1,6 @@
-package kg.peaksoft.ebookb4.security.jwt;
+package kg.peaksoft.ebookb4.config.jwt;
 
-import kg.peaksoft.ebookb4.security.services.UserDetailsServiceImpl;
+import kg.peaksoft.ebookb4.config.services.UserDetailsServiceImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
